+++ srcs/http/method.cpp
@@ -3,7 +3,7 @@
 #include "const_html_filename.hpp"
 #include "const_response_key_map.hpp"
 #include "const_status_phrase.hpp"
-#include "util/util.hpp"
+#include "utils/util.hpp"
 #include <string>
 #include <unistd.h>
