+++ srcs/utils/parser_utils.hpp
@@ -1,0 +1,9 @@
+#ifndef SRCS_UTILS_PARSER_UTILS_HPP
+#define SRCS_UTILS_PARSER_UTILS_HPP
+
+#include <string>
+
+bool is_ip(const std::string &token);
+bool is_digits(const std::string &token);
+
+#endif /* SRCS_UTILS_PARSER_UTILS_HPP */
