+++ srcs/utils/util.hpp
@@ -1,5 +1,5 @@
-#ifndef SRCS_UTIL_HPP
-#define SRCS_UTIL_HPP
+#ifndef SRCS_UTILS_UTIL_HPP
+#define SRCS_UTILS_UTIL_HPP
 
 #include <cerrno>
 #include <cstring>
@@ -17,10 +17,6 @@ bool is_file_exists(const char *path);
 std::string tostring(const size_t val);
 std::string read_file_tostring(const char *file_path);
 
-bool        is_ip(const std::string &token);
-bool        is_uint8(const std::string &token);
-bool        is_digits(const std::string &token);
-
 template <typename T>
 std::string to_string(T val) {
   std::ostringstream ss;
@@ -28,4 +24,4 @@ std::string to_string(T val) {
   return ss.str();
 }
 
-#endif /* SRCS_UTIL_HPP */
+#endif /* SRCS_UTILS_UTIL_HPP */
